src/requirements.txt
opencv-python
opencv-contrib-python
prettytable
tqdm
yacs
torchfile
plyfile
munkres
chumpy
trimesh
pyrender
shyaml
numpy-quaternion